+++ components/merino/Cargo.toml
@@ -7,8 +7,6 @@ edition = "2021"
 uniffi = { version = "0.28.2" }
 serde = { version = "1", features=["derive"] }
 serde_json = "1"
-uuid = {version = "1.12.1", features=["serde"]}
-url = {version = "2.5.4", features=["serde"]}
 
 [build-dependencies]
 uniffi = { version = "0.28.2", features = ["build"] }
